refactor: move Conversions under Tool namespace
Reviewer::Conversions → Reviewer::Tool::Conversions since the module
only converts to Tool. Delete old lib/reviewer/conversions.rb.
Remove unused Conversions include from Command::String.
Reorder requires in reviewer.rb so Tool loads before Command.
