+++ lib/reviewer.rb
@@ -5,25 +5,24 @@ require 'forwardable'
 # require 'logger'
 require 'rainbow'
 
-require_relative 'reviewer/conversions'
+require_relative 'reviewer/configuration'
+require_relative 'reviewer/history'
+require_relative 'reviewer/loader'
+require_relative 'reviewer/tool'
+require_relative 'reviewer/tools'
 
 require_relative 'reviewer/arguments'
 require_relative 'reviewer/batch'
 require_relative 'reviewer/capabilities'
 require_relative 'reviewer/command'
-require_relative 'reviewer/configuration'
 require_relative 'reviewer/failed_files'
 require_relative 'reviewer/guidance'
-require_relative 'reviewer/history'
 require_relative 'reviewer/keywords'
-require_relative 'reviewer/loader'
 require_relative 'reviewer/output'
 require_relative 'reviewer/prompt'
 require_relative 'reviewer/report'
 require_relative 'reviewer/runner'
 require_relative 'reviewer/shell'
-require_relative 'reviewer/tool'
-require_relative 'reviewer/tools'
 require_relative 'reviewer/setup'
 require_relative 'reviewer/doctor'
 require_relative 'reviewer/version'
